Specs & refactoring

+++ lib/core/types.rb
@@ -12,7 +12,7 @@ module Core
     SequelDataset = Types.Instance(Sequel::Dataset)
     Entity = Types::Nominal::Class
 
-    Latitude = Params::Decimal.constrained(included_in: COORDINATES_RANGE[:lat])
-    Longitude = Params::Decimal.constrained(included_in: COORDINATES_RANGE[:lng])
+    Latitude = Params::Float.constrained(included_in: COORDINATES_RANGE[:lat])
+    Longitude = Params::Float.constrained(included_in: COORDINATES_RANGE[:lng])
   end
 end
